# swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Phonepe Backend",
        "version": "1.0.0",
        "description": "Phonepe Backend API Documentation"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "summary": "backend health checkpoint",
                "responses": {
                    "200": {
                        "description": "server is running"
                    },
                    "500": {
                        "description": "server is not running"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "9876543210"
                                    }
                                },
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "phone"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        },
                                        "upiId": {
                                            "type": "string"
                                        },
                                        "balance": {
                                            "type": "number"
                                        },
                                        "hasMpinSet": {
                                            "type": "boolean"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (missing fields or user exists)"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}